Switched from double to float. Switched from C-casts to static_cast<>. Changed method for triangle intersection determination from areas-based to Möller–Trumbore intersection algorithm

+++ SPTracer/src/Color/SRGB.cpp
@@ -8,15 +8,15 @@ namespace SPTracer
 	Vec3 SRGB::GetRGB(const Vec3 & xyz) const
 	{
 		return Vec3{
-			GammaCorrect( 3.2406 * xyz.x - 1.5372 * xyz.y - 0.4986 * xyz.z),	// r
-			GammaCorrect(-0.9689 * xyz.x + 1.8758 * xyz.y + 0.0415 * xyz.z),	// g
-			GammaCorrect( 0.0557 * xyz.x - 0.2040 * xyz.y + 1.0570 * xyz.z)		// b
+			GammaCorrect( 3.2406f * xyz.x - 1.5372f * xyz.y - 0.4986f * xyz.z),		// r
+			GammaCorrect(-0.9689f * xyz.x + 1.8758f * xyz.y + 0.0415f * xyz.z),		// g
+			GammaCorrect( 0.0557f * xyz.x - 0.2040f * xyz.y + 1.0570f * xyz.z)		// b
 		};
 	}
 
-	double SRGB::GammaCorrect(double c) const
+	float SRGB::GammaCorrect(float c) const
 	{
-		return c <= 0.0031308 ? 12.92 * c : 1.055 * std::pow(c, 1.0 / 2.4) - 0.055;
+		return c <= 0.0031308f ? 12.92f * c : 1.055f * std::pow(c, 1.0f / 2.4f) - 0.055f;
 	}
 
 }
