+++ SPTracer/src/Model/Material/LambertianMaterial.h
@@ -15,12 +15,12 @@ namespace SPTracer
 	public:
 		LambertianMaterial(std::unique_ptr<Color> diffuseReflactance);
 		
-		bool GetNewRay(const Ray& ray, const Intersection& intersection, double waveLength, Ray& newRay, WeightFactors& weightFactors) const override;
+		bool GetNewRay(const Ray& ray, const Intersection& intersection, float waveLength, Ray& newRay, WeightFactors& weightFactors) const override;
 		bool IsEmissive() const override;
-		double GetRadiance(const Ray& ray, const Intersection& intersection, double waveLength) const override;
+		float GetRadiance(const Ray& ray, const Intersection& intersection, float waveLength) const override;
 
 	private:
-		double absorptionProbability_;
+		float absorptionProbability_;
 		std::unique_ptr<Color> diffuseReflactance_;
 	};
