+++ SPTracer/src/Model/Material/Material.h
@@ -16,9 +16,9 @@ namespace SPTracer
 	public:
 		virtual ~Material() { };
 
-		virtual bool GetNewRay(const Ray& ray, const Intersection& intersection, double waveLength, Ray& newRay, WeightFactors& weightFactors) const = 0;
+		virtual bool GetNewRay(const Ray& ray, const Intersection& intersection, float waveLength, Ray& newRay, WeightFactors& weightFactors) const = 0;
 		virtual bool IsEmissive() const = 0;
-		virtual double GetRadiance(const Ray& ray, const Intersection& intersection, double waveLength) const = 0;
+		virtual float GetRadiance(const Ray& ray, const Intersection& intersection, float waveLength) const = 0;
 
 	protected:
 		Material() { };
